refactor options, show labels and markers
